+++ NEWS.md
@@ -1,5 +1,7 @@
 # ggplot2 (development version)
 
+* When discrete breaks have names, they'll be used as labels by default 
+  (@teunbrand, #6147).
 * The helper function `is.waiver()` is now exported to help extensions to work
   with `waiver()` objects (@arcresu, #6173).
 * Date(time) scales now throw appropriate errors when `date_breaks`,